workers/chat-export-worker/requirements.txt
aio-pika==9.4.0
aiohttp==3.9.5
azure-storage-blob==12.19.1
bleach==6.1.0
Jinja2==3.1.3
prometheus-client==0.20.0
psutil==5.9.8
pydantic==2.6.3
redis==5.0.3
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
weasyprint==61.2
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-aiohttp-client==0.45b0
